% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troca.R
\name{troca}
\alias{troca}
\title{Create the TROCA shiny app}
\usage{
troca(chat, storeName = "troca")
}
\arguments{
\item{chat}{Ellmer chat.}

\item{storeName}{Name of the store or path to it, if it does not exist it
will be created.}
}
\value{
Opens the shiny app
}
\description{
Create the TROCA shiny app
}
