% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troca.R
\name{troca}
\alias{troca}
\title{Create the TROCA shiny app}
\usage{
troca(store, chat)
}
\arguments{
\item{store}{Link to the store object.}

\item{chat}{Ellmer chat.}
}
\value{
Opens the shiny app
}
\description{
Create the TROCA shiny app
}
